GP-29009 Determine on hold status based on new "status" property
This aligns the import processor with the latest API change scheduled
for May 10th which will remove "on_hold" and instead use a "status"
property to determine whether a donation/petition record is on hold.

The listed status values related to Dataflow (i.e. anything but "none")
may be used incorrectly, as we don't use the feature and have no way
to test that.