GP-40363 Add Order Type mapping
This allows webshop order types to be configured and mapped
based on the membership type, with a default fallback option.

It also adds support for an optional order_type field in the
Formunauts donation payload which would allow other order types
to be used, e.g. based on form input.